+++ continuousintegration/docker/ctafrontend/cc7/opt/run/bin/ctacli.sh
@@ -17,13 +17,11 @@
 
 . /opt/run/bin/init_pod.sh
 
-if [ ! -e /etc/buildtreeRunner ]; then
-  yum-config-manager --enable cta-artifacts
-  yum-config-manager --enable ceph
+yum-config-manager --enable cta-artifacts
+yum-config-manager --enable ceph
 
-  # Install missing RPMs
-  yum -y install cta-cli cta-debuginfo
-fi
+# Install missing RPMs
+yum -y install cta-cli cta-debuginfo
 
 cat <<EOF > /etc/cta/cta-cli.conf
 # The CTA frontend address in the form <FQDN>:<TCPPort>