+++ continuousintegration/docker/ctafrontend/cc7/opt/run/bin/init_pod.sh
@@ -61,27 +61,14 @@ if [ "-$(cat /proc/1/cmdline 2>&1 | sed -e 's/\x0//g;s/init.*/init/')-" == '-/us
   CI_CONTEXT="systemd"
 fi
 
+# some yum optimisations for the standard system
+SQUID_PROXY=squid.kube-system.svc.cluster.local
+ping -W 1 -c1 ${SQUID_PROXY} &>/dev/null && yum() { echo "Using SQUID proxy ${SQUID_PROXY}"; http_proxy=${SQUID_PROXY}:3128 /usr/bin/yum $@; }
 
-# redefining yum commands if in BUILDTREE environment
-if [[ -n ${BUILDTREE_BASE} ]]; then
-  echo "Configuring BUILDTREE environment"
-  yum-config-manager() { echo "Skipping yum-config-manager $@"; }
-  yum() { echo "Skipping yum $@"; }
-  /opt/run/bin/mkSymlinks.sh
+if test -f "/etc/config/eos/eos4"; then
+  # Switch to EOS-4 versionlock
+  /opt/run/bin/cta-versionlock --file /etc/yum/pluginconf.d/versionlock.list config eos4
 
-  # cta:tape needed for ctafrontend and taped but adding it inconditionally in buildtree
-  echo "Adding cta user and group"
-  /usr/bin/getent passwd cta || /usr/sbin/useradd -s /bin/nologin -c "CTA system account" -g tape cta
-  ## The following is not working as one cannot refresh groups in current shell...
-  # echo 'Adding cta in tape group (grant access to /dev/sg* /dev/st* /dev/nst* devices'
-  # /usr/sbin/usermod -a -G tape cta
-  # Chmod is the only way to go
-  chmod 666 /dev/nst* /dev/st* /dev/sg*
-  # creating /var/log/cta needed by taped
-  mkdir -p /var/log/cta
-  chmod a+rwx /var/log/cta
-else
-  # some yum optimisations for the standard system
-  SQUID_PROXY=squid.kube-system.svc.cluster.local
-  ping -W 1 -c1 ${SQUID_PROXY} &>/dev/null && yum() { echo "Using SQUID proxy ${SQUID_PROXY}"; http_proxy=${SQUID_PROXY}:3128 /usr/bin/yum $@; }
+  yum-config-manager --disable cta-ci-eos-5
+  yum-config-manager --enable cta-ci-eos
 fi